pyproject.toml
[project]
name = "swapi-scraper"
version = "0.1.0"
description = "SWAPI Scraper with FastAPI interface and Google Cloud Storage integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.2",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "uvicorn[standard]>=0.38.0",
    "google-cloud-storage>=2.13.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy",
    "psycopg2-binary",
    "backoff"
]
